completed middleware
